(test) Add fitting GLM model and creation of a new GLM model
