+++ src/main/kotlin/org/jetbrains/bio/statistics/mixture/ZeroPoissonMixture.kt
@@ -2,17 +2,24 @@ package org.jetbrains.bio.statistics.mixture
 import org.apache.commons.math3.linear.RealMatrix
 import org.apache.commons.math3.linear.RealVector
 import org.apache.log4j.Level
+import org.jetbrains.bio.big.BigWigFile
+import org.jetbrains.bio.coverage.Coverage
 import org.jetbrains.bio.dataframe.DataFrame
 import org.jetbrains.bio.genome.Chromosome
+import org.jetbrains.bio.genome.ChromosomeRange
 import org.jetbrains.bio.genome.Genome
 import org.jetbrains.bio.genome.GenomeQuery
+import org.jetbrains.bio.genome.sequence.TwoBitSequence
 import org.jetbrains.bio.query.ReadsQuery
 import org.jetbrains.bio.statistics.Fitter
 import org.jetbrains.bio.statistics.Preprocessed
-import org.jetbrains.bio.statistics.emission.*
+import org.jetbrains.bio.statistics.emission.ConstantIntegerEmissionScheme
+import org.jetbrains.bio.statistics.emission.EmissionScheme
+import org.jetbrains.bio.statistics.emission.PoissonRegressionEmissionScheme
 import org.jetbrains.bio.util.Logs
 import org.jetbrains.bio.viktor.F64Array
 import org.jetbrains.bio.viktor.asF64Array
+import java.nio.file.Path
 import java.nio.file.Paths
 import java.util.function.IntPredicate
 import kotlin.math.ln
@@ -86,10 +93,57 @@ class ZeroPoissonMixture(
                 check(emissions.isNotEmpty()) { "Model can't be trained on empty coverage, exiting." }
                 return ZeroPoissonMixture(
                         doubleArrayOf(1 / 3.0, 1 / 3.0, 1 / 3.0).asF64Array(),
-                        listOf("x1", "x2", "x3", "x4"),
+                        listOf("input", "GC", "mappability", "GC2"),
                         arrayOf(doubleArrayOf(0.0, 0.0, 0.0, 0.0, 0.0), doubleArrayOf(1.0, 0.0, 0.0, 0.0, 0.0)))
             }
         }
+
+        fun getIntCover(chr1: Chromosome, coverage: Coverage, bin: Int): IntArray {
+            val len = (chr1.length - 1) / bin + 1
+            val cover = IntArray(len)
+            for (i in 0 until len - 1) {
+                cover[i] = coverage.getBothStrandsCoverage(ChromosomeRange(i * bin, (i + 1) * bin, chr1))
+            }
+            cover[len - 1] = coverage.getBothStrandsCoverage(ChromosomeRange((len-1) * bin, chr1.length, chr1))
+            return cover
+        }
+        fun getDoubleCover(chr1: Chromosome, coverage: Coverage, bin: Int): DoubleArray {
+            val len = (chr1.length - 1) / bin + 1
+            val cover = DoubleArray(len)
+            for (i in 0 until len - 1) {
+                cover[i] = coverage
+                        .getBothStrandsCoverage(ChromosomeRange(i * bin, (i + 1) * bin, chr1))
+                        .toDouble()
+            }
+            cover[len - 1] = coverage
+                    .getBothStrandsCoverage(ChromosomeRange((len-1) * bin, chr1.length, chr1))
+                    .toDouble()
+            return cover
+        }
+        fun getGC(chr1: Chromosome, bin: Int): DoubleArray {
+            val len = (chr1.length - 1) / bin + 1
+            val seq: TwoBitSequence = chr1.sequence
+            val GCcontent = DoubleArray(len)
+            for (i in 0 until len - 1) {
+                GCcontent[i] = seq.substring(i*bin, (i + 1)*bin).count { it == 'c' || it == 'g' }.toDouble()/bin
+            }
+            GCcontent[len - 1] = seq
+                    .substring((len-1)*bin, seq.length)
+                    .count { it == 'c'|| it == 'g' }
+                    .toDouble()/( seq.length - (len-1)*bin)
+            return GCcontent
+        }
+
+        fun getMappability(chr1: Chromosome, path_mappability: Path, bin: Int): DoubleArray {
+            val mapSummary = BigWigFile
+                    .read(path_mappability)
+                    .summarize(chr1.name, 0, chr1.length - chr1.length%bin, numBins = (chr1.length - 1)/bin)
+            val result = DoubleArray (mapSummary.size + 1) {
+                if (it < mapSummary.size) mapSummary[it].sum/bin else 1.0}
+            result[mapSummary.size] = BigWigFile
+                    .read(path_mappability)
+                    .summarize(chr1.name, chr1.length - chr1.length%bin, 0)[0].sum / chr1.length%bin
+            return result}
 
         fun fit(
                 dirIn: String,
@@ -130,14 +184,14 @@ class ZeroPoissonMixture(
 
             val covar = DataFrame()
                     .with("y", coverMe)
-                    .with("x1", coverInput)
-                    .with("x2", GCcontent)
-                    .with("x3", mappability)
-                    .with("x4", DoubleArray(GCcontent.size) {GCcontent[it]*GCcontent[it]})
+                    .with("input", coverInput)
+                    .with("GC", GCcontent)
+                    .with("mappability", mappability)
+                    .with("GC2", DoubleArray(GCcontent.size) {GCcontent[it]*GCcontent[it]})
             Logs.addConsoleAppender(Level.DEBUG)
             val yaMix = ZeroPoissonMixture(
                     doubleArrayOf(1 / 3.0, 1 / 3.0, 1 / 3.0).asF64Array(),
-                    listOf("x1", "x2", "x3", "x4"),
+                    listOf("input", "GC", "mappability", "GC2"),
                     arrayOf(doubleArrayOf(0.0, 0.0, 0.0, 0.0, 0.0), doubleArrayOf(1.0, 0.0, 0.0, 0.0, 0.0)))
             yaMix.fit(Preprocessed.of(covar), 1e-3, 2)
             bic[index] = yaMix.BIC(covar)
